results/final_count/run-20230926-130438_LLMatic-zero_shot_donnemartin-system-design-oop-whatsapp-donnemartin-system-design-oop-url_shortener-donnemartin-system-design-oop-dropbox-donnemartin-system-design-oop-twitter_0_10-runs_log/dropbox/5/LLMatic/test_user_management.py
import user_management

def test_register_user():
	# Test with valid data
	assert user_management.register_user('John Doe', 'john.doe@example.com', 'password123') == True
	assert 'john.doe@example.com' in user_management.users_db

	# Test with invalid email
	assert user_management.register_user('John Doe', 'john.doe', 'password123') == False

	# Test with weak password
	assert user_management.register_user('John Doe', 'john.doe@example.com', 'password') == False

def test_authenticate_user():
	# Test with valid data
	user_management.register_user('Jane Doe', 'jane.doe@example.com', 'password123')
	assert user_management.authenticate_user('jane.doe@example.com', 'password123') == 'Login successful'

	# Test with invalid email
	assert user_management.authenticate_user('jane.doe', 'password123') == 'Login failed'

	# Test with incorrect password
	assert user_management.authenticate_user('jane.doe@example.com', 'password') == 'Login failed'
